Extracted domain model in a separate module
