+++ domain-model/src/main/java/com/ridesharing/domain/model/ride/RideDto.java
@@ -1,12 +1,9 @@
-package ride.sharing.app.rideplannerservice.domain;
+package com.ridesharing.domain.model.ride;
 
-import com.ridesharing.domain.model.ride.RideStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.relational.core.mapping.Column;
 
 import java.io.Serializable;
 
@@ -14,17 +11,11 @@ import java.io.Serializable;
 @NoArgsConstructor
 @Builder
 @Data
-public class Ride implements Cloneable, Serializable {
-    @Id
-    private Long id;
-
-    @Column("pickup_location")
+public class RideDto implements Cloneable, Serializable  {
     private String pickupLocation;
 
-    @Column("destination")
     private String destination;
 
-    @Column("ride_status")
     private RideStatus rideStatus;
 
     @Override
