+++ pom.xml
@@ -1,76 +1,78 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+      
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.springframework.boot</groupId>
+    <artifactId>spring-boot-starter-parent</artifactId>
+    <version>2.4.3</version>
+    <relativePath/>
+    <!-- lookup parent from repository -->
+  </parent>
+      
+  <groupId>com.ridesharing</groupId>
+  <artifactId>ride-sharing-root</artifactId>
+  <version>1.0-SNAPSHOT</version>
+      
+  <name>ride-sharing-root</name>
+      
+  <properties>
+    <revision>1.0-SNAPSHOT</revision>
 
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.4.3</version>
-        <relativePath/>
-        <!-- lookup parent from repository -->
-    </parent>
-
-    <groupId>com.ridesharing</groupId>
-    <artifactId>ride-sharing-root</artifactId>
-    <version>1.0-SNAPSHOT</version>
-
-    <name>ride-sharing-root</name>
-
-    <properties>
-        <revision>1.0-SNAPSHOT</revision>
-
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-
-        <java.version>11</java.version>
-        <maven.compiler.source>${java.version}</maven.compiler.source>
-        <maven.compiler.target>${java.version}</maven.compiler.target>
-
-        <spring-cloud.version>2020.0.1</spring-cloud.version>
-        <model.mapper.version>2.3.9</model.mapper.version>
-    </properties>
-
-    <packaging>pom</packaging>
-
-    <build>
-        <pluginManagement>
-            <plugins>
-                <plugin>
-                    <groupId>org.springframework.boot</groupId>
-                    <artifactId>spring-boot-maven-plugin</artifactId>
-                    <configuration>
-                        <excludes>
-                            <exclude>
-                                <groupId>org.projectlombok</groupId>
-                                <artifactId>lombok</artifactId>
-                            </exclude>
-                        </excludes>
-                    </configuration>
-                </plugin>
-            </plugins>
-        </pluginManagement>
-    </build>
-
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>org.springframework.cloud</groupId>
-                <artifactId>spring-cloud-dependencies</artifactId>
-                <version>${spring-cloud.version}</version>
-                <type>pom</type>
-                <scope>import</scope>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
-
-    <modules>
-        <module>config-service</module>
-        <module>eureka-service</module>
-        <module>gateway-service</module>
-        <module>ride-planner-service</module>
-        <module>gps-service</module>
-        <module>dynamic-delta-billing-service</module>
-    </modules>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+            
+    <java.version>11</java.version>
+    <maven.compiler.source>${java.version}</maven.compiler.source>
+    <maven.compiler.target>${java.version}</maven.compiler.target>
+            
+    <spring-cloud.version>2020.0.1</spring-cloud.version>
+            
+    <model.mapper.version>2.3.9</model.mapper.version>
+  </properties>
+      
+  <packaging>pom</packaging>
+      
+  <build>
+    <pluginManagement>
+      <plugins>
+        <plugin>
+          <groupId>org.springframework.boot</groupId>
+          <artifactId>spring-boot-maven-plugin</artifactId>
+          <configuration>
+            <excludes>
+              <exclude>
+                <groupId>org.projectlombok</groupId>
+                <artifactId>lombok</artifactId>
+              </exclude>
+            </excludes>
+          </configuration>
+        </plugin>
+      </plugins>
+    </pluginManagement>
+  </build>
+      
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>org.springframework.cloud</groupId>
+        <artifactId>spring-cloud-dependencies</artifactId>
+        <version>${spring-cloud.version}</version>
+        <type>pom</type>
+        <scope>import</scope>
+      </dependency>
+    </dependencies>
+        
+  </dependencyManagement>
+      
+  <modules>
+    <module>config-service</module>
+    <module>eureka-service</module>
+    <module>gateway-service</module>
+    <module>ride-planner-service</module>
+    <module>gps-service</module>
+    <module>dynamic-delta-billing-service</module>
+    <module>domain-model</module>
+  </modules>
+  
 </project>
