+++ ride-planner-service/pom.xml
@@ -15,6 +15,12 @@
     <dependencies>
         <!-- Spring backbone dependencies -->
         <dependency>
+            <groupId>com.ridesharing</groupId>
+            <artifactId>domain-model</artifactId>
+            <version>${project.version}</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-webflux</artifactId>
         </dependency>
