+++ ride-planner-service/src/main/java/ride/sharing/app/rideplannerservice/dao/service/RidePlanningServiceImpl.java
@@ -1,5 +1,8 @@
 package ride.sharing.app.rideplannerservice.dao.service;
 
+import com.ridesharing.domain.model.ride.RideDto;
+import com.ridesharing.domain.model.ride.events.EventType;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import reactor.core.publisher.Flux;
@@ -7,7 +10,6 @@ import reactor.core.publisher.Mono;
 import ride.sharing.app.rideplannerservice.dao.RideDAO;
 import ride.sharing.app.rideplannerservice.domain.Ride;
 import ride.sharing.app.rideplannerservice.domain.RideRequest;
-import ride.sharing.app.rideplannerservice.events.EventType;
 import ride.sharing.app.rideplannerservice.eventsender.EventSender;
 
 @Service
@@ -16,16 +18,19 @@ public class RidePlanningServiceImpl implements RidePlanningService {
 
     private final EventSender eventSender;
 
+    private ModelMapper modelMapper;
+
     public RidePlanningServiceImpl(RideDAO rideDAO, EventSender eventSender) {
         this.rideDAO = rideDAO;
         this.eventSender = eventSender;
+        this.modelMapper = new ModelMapper();
     }
 
     @Transactional
     @Override
     public Mono<Ride> createRide(RideRequest rideRequest) {
         return rideDAO.createRide(rideRequest)
-                .doOnNext(ride -> eventSender.send(EventType.RIDE_CREATED, ride));
+                .doOnNext(ride -> eventSender.send(EventType.RIDE_CREATED, getDto(ride)));
     }
 
     @Override
@@ -37,6 +42,10 @@ public class RidePlanningServiceImpl implements RidePlanningService {
     @Override
     public Mono<Ride> updateRide(Long rideId, RideRequest updateRideRequest) {
         return rideDAO.updateRide(rideId, updateRideRequest)
-                .doOnNext(ride -> eventSender.send(EventType.RIDE_UPDATED, ride));
+                .doOnNext(ride -> eventSender.send(EventType.RIDE_UPDATED, getDto(ride)));
+    }
+
+    private RideDto getDto(Ride ride) {
+        return modelMapper.map(ride, RideDto.class);
     }
 }
