+++ ride-planner-service/src/main/java/ride/sharing/app/rideplannerservice/eventsender/EventSender.java
@@ -1,8 +1,8 @@
 package ride.sharing.app.rideplannerservice.eventsender;
 
-import ride.sharing.app.rideplannerservice.domain.Ride;
-import ride.sharing.app.rideplannerservice.events.EventType;
+import com.ridesharing.domain.model.ride.RideDto;
+import com.ridesharing.domain.model.ride.events.EventType;
 
 public interface EventSender {
-    void send(EventType eventType, Ride ride);
+    void send(EventType eventType, RideDto ride);
 }
