+++ ride-planner-service/src/test/java/ride/sharing/app/rideplannerservice/controller/RidePlanningControllerTest.java
@@ -1,5 +1,6 @@
 package ride.sharing.app.rideplannerservice.controller;
 
+import com.ridesharing.domain.model.ride.RideUpdateType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,8 +13,9 @@ import org.springframework.test.web.reactive.server.WebTestClient;
 import org.springframework.web.reactive.function.BodyInserters;
 import ride.sharing.app.rideplannerservice.data.TestDatabaseInitializer;
 import ride.sharing.app.rideplannerservice.domain.Ride;
-import ride.sharing.app.rideplannerservice.domain.enums.RideUpdateType;
 
+import static com.ridesharing.domain.model.ride.RideStatus.CANCELLED_BY_CLIENT;
+import static com.ridesharing.domain.model.ride.RideStatus.NEW;
 import static ride.sharing.app.rideplannerservice.data.TestConstants.DESTINATION;
 import static ride.sharing.app.rideplannerservice.data.TestConstants.PICKUP_LOCATION;
 import static ride.sharing.app.rideplannerservice.data.TestConstants.UPDATED_DESTINATION;
@@ -21,8 +23,6 @@ import static ride.sharing.app.rideplannerservice.data.TestConstants.UPDATED_PIC
 import static ride.sharing.app.rideplannerservice.data.TestDataProvider.createRideEntity;
 import static ride.sharing.app.rideplannerservice.data.TestDataProvider.createRideRequest;
 import static ride.sharing.app.rideplannerservice.data.TestDatabaseInitializer.insertIntoDatabase;
-import static ride.sharing.app.rideplannerservice.domain.enums.RideStatus.CANCELLED_BY_CLIENT;
-import static ride.sharing.app.rideplannerservice.domain.enums.RideStatus.NEW;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class RidePlanningControllerTest {
